requirements.txt
Crawl4AI==0.4.247
python-dotenv==1.0.1
pydantic==2.10.6
selenium==4.18.1
webdriver-manager==4.0.1
